Moved eval for scope
